macros: Merge ruma-identifiers-macros into ruma-macros
