--- crates/ruma-identifiers-macros/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-description = "Procedural macros for creating Matrix identifiers."
-homepage = "https://www.ruma.io/"
-keywords = ["matrix", "chat", "messaging", "ruma"]
-license = "MIT"
-name = "ruma-identifiers-macros"
-repository = "https://github.com/ruma/ruma"
-version = "0.22.0"
-edition = "2018"
-
-[lib]
-proc-macro = true
-
-[dependencies]
-quote = "1.0.8"
-ruma-identifiers-validation = { version = "0.7.0", path = "../ruma-identifiers-validation", default-features = false }
-syn = "1.0.55"
