+++ crates/ruma-identifiers/Cargo.toml
@@ -24,8 +24,8 @@ serde = ["ruma-serde", "serde1"]
 either = { version = "1.6.1", optional = true }
 percent-encoding = "2.1.0"
 rand_crate = { package = "rand", version = "0.8.3", optional = true }
-ruma-identifiers-macros = { version = "=0.22.0", path = "../ruma-identifiers-macros" }
 ruma-identifiers-validation = { version = "0.7.0", path = "../ruma-identifiers-validation", default-features = false }
+ruma-macros = { version = "=0.1.0", path = "../ruma-macros" }
 ruma-serde = { version = "0.6.0", path = "../ruma-serde", optional = true }
 ruma-serde-macros = { version = "0.6.0", path = "../ruma-serde-macros" }
 # Renamed so we can have a serde feature.
